Broke out Wallet and Nodes components
