+++ src/main/java/org/unigrid/janus/controller/view/NodesController.java
@@ -1,0 +1,35 @@
+/*
+	The Janus Wallet
+	Copyright © 2021 The Unigrid Foundation
+
+	This program is free software: you can redistribute it and/or modify it under the terms of the
+	addended GNU Affero General Public License as published by the Free Software Foundation, version 3
+	of the License (see COPYING and COPYING.addendum).
+
+	This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+	even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+	GNU Affero General Public License for more details.
+
+	You should have received an addended copy of the GNU Affero General Public License with this program.
+	If not, see <http://www.gnu.org/licenses/> and <https://github.com/unigrid-project/janus-java>.
+*/
+
+package org.unigrid.janus.controller.view;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.fxml.Initializable;
+import org.unigrid.janus.model.service.DebugService;
+import org.unigrid.janus.model.service.RPCService;
+import org.unigrid.janus.model.service.WindowService;
+
+public class NodesController implements Initializable {
+	private static DebugService debug = new DebugService();
+	private static RPCService rpc = new RPCService();
+	private static WindowService window = new WindowService();
+
+	@Override
+	public void initialize(URL url, ResourceBundle rb) {
+		/* Empty on purpose */
+	}
+}
